feat(config): Rename `basic_auth` to `auth`
